All javascript  is now loaded using AMD
